Gear now modifies player stats

+++ Assets/Scripts/UnitStats.cs
@@ -6,33 +6,33 @@ namespace Assets.Scripts
     {
         public int Id { get; set; }
         public string Name { get; set; }
-        public int MaxHealth { get; set; }
-        public int MaxMana { get; set; }
-        public int AttackPower { get; set; }
-        public int MagicPower { get; set; }
+        public float MaxHealth { get; set; }
+        public float MaxMana { get; set; }
+        public float AttackPower { get; set; }
+        public float MagicPower { get; set; }
         public int Speed { get; set; }
-        public int AttackDefense { get; set; }
-        public int MagicDefense { get; set; }
-        public int Accuracy { get; set; }
-        public int Evasion { get; set; }
-        public int CritChance { get; set; }
-        public int CritMulti { get; set; }
-        public int IncCritChance { get; set; }
-        public int TypeStatusChance { get; set; }
-        public int MentalStatusChance { get; set; }
-        public int StatusPower { get; set; }
-        public int IncTypeStatus { get; set; }
-        public int IncMentalStatus { get; set; }
-        public int ResistPhysical { get; set; }
-        public int ResistProjectile { get; set; }
-        public int ResistElectric { get; set; }
-        public int ResistCold { get; set; }
-        public int ResistFire { get; set; }
-        public int ResistWind { get; set; }
-        public int ResistArcane { get; set; }
-        public int ResistPsychic { get; set; }
-        public int ResistLight { get; set; }
-        public int ResistDark { get; set; }
+        public float AttackDefense { get; set; }
+        public float MagicDefense { get; set; }
+        public float Accuracy { get; set; }
+        public float Evasion { get; set; }
+        public float CritChance { get; set; }
+        public float CritMulti { get; set; }
+        public float IncCritChance { get; set; }
+        public float TypeStatusChance { get; set; }
+        public float MentalStatusChance { get; set; }
+        public float StatusPower { get; set; }
+        public float IncTypeStatus { get; set; }
+        public float IncMentalStatus { get; set; }
+        public float ResistPhysical { get; set; }
+        public float ResistProjectile { get; set; }
+        public float ResistElectric { get; set; }
+        public float ResistCold { get; set; }
+        public float ResistFire { get; set; }
+        public float ResistWind { get; set; }
+        public float ResistArcane { get; set; }
+        public float ResistPsychic { get; set; }
+        public float ResistLight { get; set; }
+        public float ResistDark { get; set; }
         [Ignore]
         public Constants.StatusTypes Status { get; set; }
